Remove old nearby implementation Add extra specs
