+++ modules/va_facilities/spec/serializers/nearby_serializer_spec.rb
@@ -30,12 +30,12 @@ RSpec.describe VaFacilities::NearbySerializer, type: :serializer do
   end
 
   it 'returns a relationship link to its facility' do
-    fac_path = "#{VaFacilities::NearbyFacilitySerializer::BASE_PATH}/facilities/vha_#{vha_648.unique_id}"
+    fac_path = "#{VaFacilities::NearbySerializer::BASE_PATH}/facilities/vha_#{vha_648.unique_id}"
     expect(relationships['va_facilities']['links']['related']).to eq(fac_path)
   end
 
   it 'returns a relationship link to its drivetime band' do
-    band_path = "#{VaFacilities::NearbyFacilitySerializer::BASE_PATH}/drivetime_bands/#{thirty.id}"
+    band_path = "#{VaFacilities::NearbySerializer::BASE_PATH}/drivetime_bands/#{thirty.id}"
     expect(relationships['drivetime_band']['links']['related']).to eq(band_path)
   end
 end
