--- spec/support/vcr_cassettes/bing/isochrone/pdx_drive_time_3000.yml
@@ -1,119 +1,0 @@
----
-http_interactions:
-- request:
-    method: get
-    uri: https://dev.virtualearth.net/REST/v1/Locations?key=fake_key&q=9729%20SE%20222nd%20Dr%20Damascus%20OR%2097089
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      User-Agent:
-        - Faraday v0.9.2
-      Accept-Encoding:
-        - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-        - "*/*"
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Cache-Control:
-        - no-cache
-      Transfer-Encoding:
-        - chunked
-      Content-Type:
-        - application/json; charset=utf-8
-      Vary:
-        - Accept-Encoding
-      Server:
-        - Microsoft-IIS/10.0
-      X-Bm-Traceid:
-        - 796ffca0d2054776a2be4f175d2e9a26
-      X-Bm-Be-Elapsed:
-        - '137'
-      X-Msedge-Ref:
-        - 'Ref A: 93D91C4A8DDF4851949C9D84B60E8228 Ref B: BN3EDGE1006 Ref C: 2019-10-08T19:00:42Z'
-      X-Bm-Fe-Elapsed:
-        - '167'
-      X-Bm-Srv:
-        - BN000021DC
-      X-Ms-Bm-Ws-Info:
-        - '0'
-      Access-Control-Allow-Origin:
-        - "*"
-      Access-Control-Allow-Methods:
-        - POST, GET, OPTIONS
-      Access-Control-Allow-Headers:
-        - Content-Type,X-FD-Features,X-FD-FLIGHT
-      X-Aspnet-Version:
-        - 4.0.30319
-      X-Powered-By:
-        - ASP.NET
-      Date:
-        - Tue, 08 Oct 2019 19:00:41 GMT
-    body:
-      encoding: ASCII-8BIT
-      string: !binary |-
-        eyJhdXRoZW50aWNhdGlvblJlc3VsdENvZGUiOiJWYWxpZENyZWRlbnRpYWxzIiwiYnJhbmRMb2dvVXJpIjoiaHR0cDpcL1wvZGV2LnZpcnR1YWxlYXJ0aC5uZXRcL0JyYW5kaW5nXC9sb2dvX3Bvd2VyZWRfYnkucG5nIiwiY29weXJpZ2h0IjoiQ29weXJpZ2h0IMKpIDIwMTkgTWljcm9zb2Z0IGFuZCBpdHMgc3VwcGxpZXJzLiBBbGwgcmlnaHRzIHJlc2VydmVkLiBUaGlzIEFQSSBjYW5ub3QgYmUgYWNjZXNzZWQgYW5kIHRoZSBjb250ZW50IGFuZCBhbnkgcmVzdWx0cyBtYXkgbm90IGJlIHVzZWQsIHJlcHJvZHVjZWQgb3IgdHJhbnNtaXR0ZWQgaW4gYW55IG1hbm5lciB3aXRob3V0IGV4cHJlc3Mgd3JpdHRlbiBwZXJtaXNzaW9uIGZyb20gTWljcm9zb2Z0IENvcnBvcmF0aW9uLiIsInJlc291cmNlU2V0cyI6W3siZXN0aW1hdGVkVG90YWwiOjEsInJlc291cmNlcyI6W3siX190eXBlIjoiTG9jYXRpb246aHR0cDpcL1wvc2NoZW1hcy5taWNyb3NvZnQuY29tXC9zZWFyY2hcL2xvY2FsXC93c1wvcmVzdFwvdjEiLCJiYm94IjpbNDUuNDQ4MDYzMjgyNDI5MzI0LC0xMjIuNDQzMDg0MjQzODUyMTMsNDUuNDU1Nzg4NzE3NTcwNjc3LC0xMjIuNDI4NDAxNzU2MTQ3ODhdLCJuYW1lIjoiOTcyOSBTRSAyMjJuZCBEciwgRGFtYXNjdXMsIE9SIDk3MDg5IiwicG9pbnQiOnsidHlwZSI6IlBvaW50IiwiY29vcmRpbmF0ZXMiOls0NS40NTE5MjYsLTEyMi40MzU3NDNdfSwiYWRkcmVzcyI6eyJhZGRyZXNzTGluZSI6Ijk3MjkgU0UgMjIybmQgRHIiLCJhZG1pbkRpc3RyaWN0IjoiT1IiLCJhZG1pbkRpc3RyaWN0MiI6IkNsYWNrYW1hcyBDb3VudHkiLCJjb3VudHJ5UmVnaW9uIjoiVW5pdGVkIFN0YXRlcyIsImZvcm1hdHRlZEFkZHJlc3MiOiI5NzI5IFNFIDIyMm5kIERyLCBEYW1hc2N1cywgT1IgOTcwODkiLCJsb2NhbGl0eSI6IkRhbWFzY3VzIiwicG9zdGFsQ29kZSI6Ijk3MDg5In0sImNvbmZpZGVuY2UiOiJIaWdoIiwiZW50aXR5VHlwZSI6IkFkZHJlc3MiLCJnZW9jb2RlUG9pbnRzIjpbeyJ0eXBlIjoiUG9pbnQiLCJjb29yZGluYXRlcyI6WzQ1LjQ1MTkyNiwtMTIyLjQzNTc0M10sImNhbGN1bGF0aW9uTWV0aG9kIjoiUm9vZnRvcCIsInVzYWdlVHlwZXMiOlsiRGlzcGxheSJdfSx7InR5cGUiOiJQb2ludCIsImNvb3JkaW5hdGVzIjpbNDUuNDUxOTE0NzE4MDcwNiwtMTIyLjQzNDQ1ODg1MTNdLCJjYWxjdWxhdGlvbk1ldGhvZCI6IlJvb2Z0b3AiLCJ1c2FnZVR5cGVzIjpbIlJvdXRlIl19XSwibWF0Y2hDb2RlcyI6WyJHb29kIl19XX1dLCJzdGF0dXNDb2RlIjoyMDAsInN0YXR1c0Rlc2NyaXB0aW9uIjoiT0siLCJ0cmFjZUlkIjoiNzk2ZmZjYTBkMjA1NDc3NmEyYmU0ZjE3NWQyZTlhMjZ8Qk4wMDAwMjFEQ3w3LjcuMC4wfFJlZiBBOiA5M0Q5MUM0QThEREY0ODUxOTQ5QzlEODRCNjBFODIyOCBSZWYgQjogQk4zRURHRTEwMDYgUmVmIEM6IDIwMTktMTAtMDhUMTk6MDA6NDJaIn0=
-    http_version:
-  recorded_at: Tue, 08 Oct 2019 19:00:42 GMT
-- request:
-    method: get
-    uri: https://dev.virtualearth.net/REST/v1/Routes/Isochrones?dateTime=07:30:00&key=fake_key&maxtime=3000&optimize=timeWithTraffic&timeUnit=minute&waypoint=45.451926,-122.435743
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      User-Agent:
-      - Faraday v0.9.2
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-  response:
-    status:
-      code: 200
-      message: Timeout occurred.
-    headers:
-      Cache-Control:
-      - no-cache
-      Transfer-Encoding:
-      - chunked
-      Content-Type:
-      - application/json; charset=utf-8
-      Vary:
-      - Accept-Encoding
-      Server:
-      - Microsoft-IIS/10.0
-      X-Bm-Traceid:
-      - 9bae4c8818274e5dbc2f2f6ccb020f6d
-      X-Bm-Be-Elapsed:
-      - '9018'
-      X-Msedge-Ref:
-      - 'Ref A: 2E6F521BF84544C79DFDD6A3B021A072 Ref B: BN3EDGE0721 Ref C: 2019-06-06T21:15:04Z'
-      X-Bm-Fe-Elapsed:
-      - '9144'
-      X-Bm-Srv:
-      - BN12F24E2D
-      X-Ms-Bm-Ws-Info:
-      - '0'
-      Access-Control-Allow-Origin:
-      - "*"
-      Access-Control-Allow-Methods:
-      - POST, GET, OPTIONS
-      Access-Control-Allow-Headers:
-      - Content-Type
-      X-Aspnet-Version:
-      - 4.0.30319
-      X-Powered-By:
-      - ASP.NET
-      Date:
-      - Thu, 06 Jun 2019 21:15:13 GMT
-    body:
-      encoding: ASCII-8BIT
-      string: !binary |-
-        eyJhdXRoZW50aWNhdGlvblJlc3VsdENvZGUiOiJWYWxpZENyZWRlbnRpYWxzIiwiYnJhbmRMb2dvVXJpIjoiaHR0cDpcL1wvZGV2LnZpcnR1YWxlYXJ0aC5uZXRcL0JyYW5kaW5nXC9sb2dvX3Bvd2VyZWRfYnkucG5nIiwiY29weXJpZ2h0IjoiQ29weXJpZ2h0IMKpIDIwMTkgTWljcm9zb2Z0IGFuZCBpdHMgc3VwcGxpZXJzLiBBbGwgcmlnaHRzIHJlc2VydmVkLiBUaGlzIEFQSSBjYW5ub3QgYmUgYWNjZXNzZWQgYW5kIHRoZSBjb250ZW50IGFuZCBhbnkgcmVzdWx0cyBtYXkgbm90IGJlIHVzZWQsIHJlcHJvZHVjZWQgb3IgdHJhbnNtaXR0ZWQgaW4gYW55IG1hbm5lciB3aXRob3V0IGV4cHJlc3Mgd3JpdHRlbiBwZXJtaXNzaW9uIGZyb20gTWljcm9zb2Z0IENvcnBvcmF0aW9uLiIsImVycm9ycyI6W3siZXJyb3JDb2RlIjoiIiwiZXJyb3JEZXRhaWxzIjpbIlRpbWVvdXQgb2NjdXJyZWQuIl19XSwicmVzb3VyY2VTZXRzIjpbXSwic3RhdHVzQ29kZSI6MjAwLCJzdGF0dXNEZXNjcmlwdGlvbiI6IlRpbWVvdXQgb2NjdXJyZWQuIiwidHJhY2VJZCI6IjliYWU0Yzg4MTgyNzRlNWRiYzJmMmY2Y2NiMDIwZjZkfEJOMTJGMjRFMkR8Ny43LjAuMHxSZWYgQTogMkU2RjUyMUJGODQ1NDRDNzlERkRENkEzQjAyMUEwNzIgUmVmIEI6IEJOM0VER0UwNzIxIFJlZiBDOiAyMDE5LTA2LTA2VDIxOjE1OjA0WiJ9
-    http_version: 
-  recorded_at: Thu, 06 Jun 2019 21:15:14 GMT
-recorded_with: VCR 3.0.3
