--- spec/support/vcr_cassettes/bing/isochrone/pdx_malformed_address.yml
@@ -1,119 +1,0 @@
----
-http_interactions:
-- request:
-    method: get
-    uri: https://dev.virtualearth.net/REST/v1/Locations?key=fake_key&q=9729%20Sbleepblap%20Damascus%20OR%2097089
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      User-Agent:
-        - Faraday v0.9.2
-      Accept-Encoding:
-        - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-        - "*/*"
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Cache-Control:
-        - no-cache
-      Transfer-Encoding:
-        - chunked
-      Content-Type:
-        - application/json; charset=utf-8
-      Vary:
-        - Accept-Encoding
-      Server:
-        - Microsoft-IIS/10.0
-      X-Bm-Traceid:
-        - 796ffca0d2054776a2be4f175d2e9a26
-      X-Bm-Be-Elapsed:
-        - '137'
-      X-Msedge-Ref:
-        - 'Ref A: 93D91C4A8DDF4851949C9D84B60E8228 Ref B: BN3EDGE1006 Ref C: 2019-10-08T19:00:42Z'
-      X-Bm-Fe-Elapsed:
-        - '167'
-      X-Bm-Srv:
-        - BN000021DC
-      X-Ms-Bm-Ws-Info:
-        - '0'
-      Access-Control-Allow-Origin:
-        - "*"
-      Access-Control-Allow-Methods:
-        - POST, GET, OPTIONS
-      Access-Control-Allow-Headers:
-        - Content-Type,X-FD-Features,X-FD-FLIGHT
-      X-Aspnet-Version:
-        - 4.0.30319
-      X-Powered-By:
-        - ASP.NET
-      Date:
-        - Tue, 08 Oct 2019 19:00:41 GMT
-    body:
-      encoding: ASCII-8BIT
-      string: !binary |-
-        eyJhdXRoZW50aWNhdGlvblJlc3VsdENvZGUiOiJWYWxpZENyZWRlbnRpYWxzIiwiYnJhbmRMb2dvVXJpIjoiaHR0cDpcL1wvZGV2LnZpcnR1YWxlYXJ0aC5uZXRcL0JyYW5kaW5nXC9sb2dvX3Bvd2VyZWRfYnkucG5nIiwiY29weXJpZ2h0IjoiQ29weXJpZ2h0IMKpIDIwMTkgTWljcm9zb2Z0IGFuZCBpdHMgc3VwcGxpZXJzLiBBbGwgcmlnaHRzIHJlc2VydmVkLiBUaGlzIEFQSSBjYW5ub3QgYmUgYWNjZXNzZWQgYW5kIHRoZSBjb250ZW50IGFuZCBhbnkgcmVzdWx0cyBtYXkgbm90IGJlIHVzZWQsIHJlcHJvZHVjZWQgb3IgdHJhbnNtaXR0ZWQgaW4gYW55IG1hbm5lciB3aXRob3V0IGV4cHJlc3Mgd3JpdHRlbiBwZXJtaXNzaW9uIGZyb20gTWljcm9zb2Z0IENvcnBvcmF0aW9uLiIsInJlc291cmNlU2V0cyI6W3siZXN0aW1hdGVkVG90YWwiOjEsInJlc291cmNlcyI6W3siX190eXBlIjoiTG9jYXRpb246aHR0cDpcL1wvc2NoZW1hcy5taWNyb3NvZnQuY29tXC9zZWFyY2hcL2xvY2FsXC93c1wvcmVzdFwvdjEiLCJiYm94IjpbNDUuNDQ4MDYzMjgyNDI5MzI0LC0xMjIuNDQzMDg0MjQzODUyMTMsNDUuNDU1Nzg4NzE3NTcwNjc3LC0xMjIuNDI4NDAxNzU2MTQ3ODhdLCJuYW1lIjoiOTcyOSBTRSAyMjJuZCBEciwgRGFtYXNjdXMsIE9SIDk3MDg5IiwicG9pbnQiOnsidHlwZSI6IlBvaW50IiwiY29vcmRpbmF0ZXMiOls0NS40NTE5MjYsLTEyMi40MzU3NDNdfSwiYWRkcmVzcyI6eyJhZGRyZXNzTGluZSI6Ijk3MjkgU0UgMjIybmQgRHIiLCJhZG1pbkRpc3RyaWN0IjoiT1IiLCJhZG1pbkRpc3RyaWN0MiI6IkNsYWNrYW1hcyBDb3VudHkiLCJjb3VudHJ5UmVnaW9uIjoiVW5pdGVkIFN0YXRlcyIsImZvcm1hdHRlZEFkZHJlc3MiOiI5NzI5IFNFIDIyMm5kIERyLCBEYW1hc2N1cywgT1IgOTcwODkiLCJsb2NhbGl0eSI6IkRhbWFzY3VzIiwicG9zdGFsQ29kZSI6Ijk3MDg5In0sImNvbmZpZGVuY2UiOiJIaWdoIiwiZW50aXR5VHlwZSI6IkFkZHJlc3MiLCJnZW9jb2RlUG9pbnRzIjpbeyJ0eXBlIjoiUG9pbnQiLCJjb29yZGluYXRlcyI6WzQ1LjQ1MTkyNiwtMTIyLjQzNTc0M10sImNhbGN1bGF0aW9uTWV0aG9kIjoiUm9vZnRvcCIsInVzYWdlVHlwZXMiOlsiRGlzcGxheSJdfSx7InR5cGUiOiJQb2ludCIsImNvb3JkaW5hdGVzIjpbNDUuNDUxOTE0NzE4MDcwNiwtMTIyLjQzNDQ1ODg1MTNdLCJjYWxjdWxhdGlvbk1ldGhvZCI6IlJvb2Z0b3AiLCJ1c2FnZVR5cGVzIjpbIlJvdXRlIl19XSwibWF0Y2hDb2RlcyI6WyJHb29kIl19XX1dLCJzdGF0dXNDb2RlIjoyMDAsInN0YXR1c0Rlc2NyaXB0aW9uIjoiT0siLCJ0cmFjZUlkIjoiNzk2ZmZjYTBkMjA1NDc3NmEyYmU0ZjE3NWQyZTlhMjZ8Qk4wMDAwMjFEQ3w3LjcuMC4wfFJlZiBBOiA5M0Q5MUM0QThEREY0ODUxOTQ5QzlEODRCNjBFODIyOCBSZWYgQjogQk4zRURHRTEwMDYgUmVmIEM6IDIwMTktMTAtMDhUMTk6MDA6NDJaIn0=
-    http_version:
-  recorded_at: Tue, 08 Oct 2019 19:00:42 GMT
-- request:
-    method: get
-    uri: https://dev.virtualearth.net/REST/v1/Routes/Isochrones?dateTime=07:30:00&key=fake_key&maxtime=1&optimize=timeWithTraffic&timeUnit=minute&waypoint=45.451926,-122.435743
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      User-Agent:
-      - Faraday v0.9.2
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Cache-Control:
-      - no-cache
-      Transfer-Encoding:
-      - chunked
-      Content-Type:
-      - application/json; charset=utf-8
-      Vary:
-      - Accept-Encoding
-      Server:
-      - Microsoft-IIS/10.0
-      X-Bm-Traceid:
-      - 2d0ce69e5be54c709317e6fa68bf30d7
-      X-Bm-Be-Elapsed:
-      - '112'
-      X-Msedge-Ref:
-      - 'Ref A: 55017C529F9647BBA5F3061803D50640 Ref B: BN3EDGE0717 Ref C: 2019-06-12T14:36:15Z'
-      X-Bm-Fe-Elapsed:
-      - '149'
-      X-Bm-Srv:
-      - BN12968898
-      X-Ms-Bm-Ws-Info:
-      - '0'
-      Access-Control-Allow-Origin:
-      - "*"
-      Access-Control-Allow-Methods:
-      - POST, GET, OPTIONS
-      Access-Control-Allow-Headers:
-      - Content-Type
-      X-Aspnet-Version:
-      - 4.0.30319
-      X-Powered-By:
-      - ASP.NET
-      Date:
-      - Wed, 12 Jun 2019 14:36:15 GMT
-    body:
-      encoding: ASCII-8BIT
-      string: !binary |-
-        eyJhdXRoZW50aWNhdGlvblJlc3VsdENvZGUiOiJWYWxpZENyZWRlbnRpYWxzIiwiYnJhbmRMb2dvVXJpIjoiaHR0cDpcL1wvZGV2LnZpcnR1YWxlYXJ0aC5uZXRcL0JyYW5kaW5nXC9sb2dvX3Bvd2VyZWRfYnkucG5nIiwiY29weXJpZ2h0IjoiQ29weXJpZ2h0IMKpIDIwMTkgTWljcm9zb2Z0IGFuZCBpdHMgc3VwcGxpZXJzLiBBbGwgcmlnaHRzIHJlc2VydmVkLiBUaGlzIEFQSSBjYW5ub3QgYmUgYWNjZXNzZWQgYW5kIHRoZSBjb250ZW50IGFuZCBhbnkgcmVzdWx0cyBtYXkgbm90IGJlIHVzZWQsIHJlcHJvZHVjZWQgb3IgdHJhbnNtaXR0ZWQgaW4gYW55IG1hbm5lciB3aXRob3V0IGV4cHJlc3Mgd3JpdHRlbiBwZXJtaXNzaW9uIGZyb20gTWljcm9zb2Z0IENvcnBvcmF0aW9uLiIsInJlc291cmNlU2V0cyI6W3siZXN0aW1hdGVkVG90YWwiOjEsInJlc291cmNlcyI6W3siX190eXBlIjoiSXNvY2hyb25lUmVzcG9uc2U6aHR0cDpcL1wvc2NoZW1hcy5taWNyb3NvZnQuY29tXC9zZWFyY2hcL2xvY2FsXC93c1wvcmVzdFwvdjEiLCJvcmlnaW4iOnsibGF0aXR1ZGUiOjQ1LjQwNjg0MSwibG9uZ2l0dWRlIjotMTIyLjQ5OTY3Mn0sInBvbHlnb25zIjpbeyJjb29yZGluYXRlcyI6W1tbNDUuNDA2NjU5LC0xMjIuNTAyMzg0XSxbNDUuNDA2OTAzLC0xMjIuNTAyMTE4XSxbNDUuNDA2MzQ1LC0xMjIuNTAxODUyXSxbNDUuNDA2NDEsLTEyMi41MDE1ODZdLFs0NS40MDU5NSwtMTIyLjUwMTMxOV0sWzQ1LjQwNjAyNywtMTIyLjUwMTA1M10sWzQ1LjQwNjE4LC0xMjIuNTAwNzg3XSxbNDUuNDA3MjE5LC0xMjIuNTAwNTIxXSxbNDUuNDA2NjgsLTEyMi41MDAyNTRdLFs0NS40MDY5LC0xMjIuNDk5OTg4XSxbNDUuNDA3MTQxLC0xMjIuNDk5NzIyXSxbNDUuNDA3MjYsLTEyMi40OTk0NTZdLFs0NS40MDcwMiwtMTIyLjQ5OTE4OV0sWzQ1LjQwNjk2LC0xMjIuNDk4OTIzXSxbNDUuNDA2ODY1LC0xMjIuNDk4NjU3XSxbNDUuNDA2Mjg1LC0xMjIuNDk4MzkxXSxbNDUuNDA2MTU1LC0xMjIuNDk4MTI0XSxbNDUuNDA2MTUsLTEyMi40OTc4NThdLFs0NS40MDYxNywtMTIyLjQ5NzU5Ml0sWzQ1LjQwNzkyMywtMTIyLjQ5NzMyNl0sWzQ1LjQwNTYyLC0xMjIuNDk3MzI2XSxbNDUuNDA1NTQsLTEyMi40OTc1OTJdLFs0NS40MDU1NSwtMTIyLjQ5Nzg1OF0sWzQ1LjQwNTU1NSwtMTIyLjQ5ODEyNF0sWzQ1LjQwNTU2LC0xMjIuNDk4MzkxXSxbNDUuNDA1NzMsLTEyMi40OTg2NTddLFs0NS40MDYzMiwtMTIyLjQ5ODkyM10sWzQ1LjQwNjQsLTEyMi40OTkxODldLFs0NS40MDU1NCwtMTIyLjQ5OTE4OV0sWzQ1LjQwNDczMywtMTIyLjUwMDExXSxbNDUuNDA0NTQ3LC0xMjIuNTAwMDM1XSxbNDUuNDA0MzYxLC0xMjIuNDk5OTddLFs0NS40MDQxNzYsLTEyMi40OTk2MTVdLFs0NS40MDM2NCwtMTIyLjQ5OTQ1Nl0sWzQ1LjQwMzYyNSwtMTIyLjQ5OTcyMl0sWzQ1LjQwMzYxLC0xMjIuNDk5OTg4XSxbNDUuNDAzNzY1LC0xMjIuNTAwMjU0XSxbNDUuNDA0MTQsLTEyMi41MDA1MjFdLFs0NS40MDU1MDMsLTEyMi41MDA3ODddLFs0NS40MDU0MjcsLTEyMi41MDEwNTNdLFs0NS40MDUzNSwtMTIyLjUwMTMxOV0sWzQ1LjQwNTEyLC0xMjIuNTAxNTg2XSxbNDUuNDA1NjQ2LC0xMjIuNTAxODUyXSxbNDUuNDA1NDgxLC0xMjIuNTAyMTE4XSxbNDUuNDA2MDU5LC0xMjIuNTAyMzg0XSxbNDUuNDA2NjU5LC0xMjIuNTAyMzg0XV1dfV19XX1dLCJzdGF0dXNDb2RlIjoyMDAsInN0YXR1c0Rlc2NyaXB0aW9uIjoiT0siLCJ0cmFjZUlkIjoiMmQwY2U2OWU1YmU1NGM3MDkzMTdlNmZhNjhiZjMwZDd8Qk4xMjk2ODg5OHw3LjcuMC4wfFJlZiBBOiA1NTAxN0M1MjlGOTY0N0JCQTVGMzA2MTgwM0Q1MDY0MCBSZWYgQjogQk4zRURHRTA3MTcgUmVmIEM6IDIwMTktMDYtMTJUMTQ6MzY6MTVaIn0=
-    http_version: 
-  recorded_at: Wed, 12 Jun 2019 14:36:15 GMT
-recorded_with: VCR 3.0.3
